Moves file events
