+++ .gitignore
@@ -246,3 +246,4 @@ legacy.db
 \#*\#
 .\#*
 .aider*
+/.agent-shell/
