+++ submit_ce/api/domain/event/__init__.py
@@ -13,8 +13,7 @@ Writing new events/commands
 Events/commands are implemented as classes that inherit from :class:`.Event`.
 It should:
 
-- Be a dataclass (i.e. be decorated with :func:`dataclasses.dataclass`).
-- Define (using :func:`dataclasses.field`) associated data.
+- Define associated data.
 - Implement a validation method with the signature
   ``validate(self, submission: Submission) -> None`` (see below).
 - Implement a projection method with the signature
@@ -62,6 +61,7 @@ from pytz import UTC
 from . import validators
 from .base import Event
 from .base import event_factory as make_event
+from .file import SetUploadPackage, UpdateUploadPackage
 from .flag import AddMetadataFlag, AddUserFlag, AddContentFlag, RemoveFlag, \
     AddHold, RemoveHold
 from .proposal import AddProposal, RejectProposal, AcceptProposal
@@ -787,95 +787,6 @@ class SetAuthors(Event):
         return submission
 
 
-class SetUploadPackage(Event):
-    """Set the upload workspace for this submission."""
-
-    NAME = "set the upload package"
-    NAMED = "upload package set"
-
-    identifier: str = field(default_factory=str)
-    checksum: str = field(default_factory=str)
-    uncompressed_size: int = field(default=0)
-    compressed_size: int = field(default=0)
-    source_format: SubmissionContent.Format = \
-        field(default=SubmissionContent.Format.UNKNOWN)
-
-    def model_post_init(self, *args, **kwargs) -> None:
-        """Make sure that `source_format` is an enum instance."""
-        if type(self.source_format) is str:
-            self.source_format = SubmissionContent.Format(self.source_format)
-
-    def validate(self, submission: Submission) -> None:
-        """Validate data for :class:`.SetUploadPackage`."""
-        validators.submission_is_not_finalized(self, submission)
-
-        if not self.identifier:
-            raise InvalidEvent(self, 'Missing upload ID')
-
-    def project(self, submission: Submission) -> Submission:
-        """Replace :class:`.SubmissionContent` metadata on the submission."""
-        submission.source_content = SubmissionContent(
-            checksum=self.checksum,
-            identifier=self.identifier,
-            uncompressed_size=self.uncompressed_size,
-            compressed_size=self.compressed_size,
-            source_format=self.source_format,
-        )
-        submission.submitter_confirmed_preview = False
-        return submission
-
-
-class UpdateUploadPackage(Event):
-    """Update the upload workspace on this submission."""
-
-    NAME = "update the upload package"
-    NAMED = "upload package updated"
-
-    checksum: str = field(default_factory=str)
-    uncompressed_size: int = field(default=0)
-    compressed_size: int = field(default=0)
-    source_format: SubmissionContent.Format = \
-        field(default=SubmissionContent.Format.UNKNOWN)
-
-    def model_post_init(self, *args, **kwargs) -> None:
-        """Make sure that `source_format` is an enum instance."""
-        if type(self.source_format) is str:
-            self.source_format = SubmissionContent.Format(self.source_format)
-
-    def validate(self, submission: Submission) -> None:
-        """Validate data for :class:`.SetUploadPackage`."""
-        validators.submission_is_not_finalized(self, submission)
-
-    def project(self, submission: Submission) -> Submission:
-        """Replace :class:`.SubmissionContent` metadata on the submission."""
-        assert submission.source_content is not None
-        assert self.source_format is not None
-        assert self.checksum is not None
-        assert self.uncompressed_size is not None
-        assert self.compressed_size is not None
-        submission.source_content.source_format = self.source_format
-        submission.source_content.checksum = self.checksum
-        submission.source_content.uncompressed_size = self.uncompressed_size
-        submission.source_content.compressed_size = self.compressed_size
-        submission.submitter_confirmed_preview = False
-        return submission
-
-
-class UnsetUploadPackage(Event):
-    """Unset the upload workspace for this submission."""
-
-    NAME = "unset the upload package"
-    NAMED = "upload package unset"
-
-    def validate(self, submission: Submission) -> None:
-        """Validate data for :class:`.UnsetUploadPackage`."""
-        validators.submission_is_not_finalized(self, submission)
-
-    def project(self, submission: Submission) -> Submission:
-        """Set :attr:`Submission.source_content` to None."""
-        submission.source_content = None
-        submission.submitter_confirmed_preview = False
-        return submission
 
 
 class ConfirmSourceProcessed(Event):
